+++ packages/jest-prosemirror/CHANGELOG.md
@@ -1,5 +1,20 @@
 # jest-prosemirror
 
+## 1.0.20
+
+> 2022-03-31
+
+### Patch Changes
+
+- Add support for Unicode Regexp in suggestion matching.
+
+  The change was required to support matching non-latin characters in `MentionAtomExtension` and `MentionExtension` i.e. by using `supportedCharacters: /\p{Letter}+/u` in `matchers` definition.
+
+  There is no need to update the code: changes are backwards compatible with no behavior change at all.
+
+- Updated dependencies []:
+  - @remirror/pm@1.0.15
+
 ## 1.0.19
 
 > 2022-03-08