+++ packages/remirror__extension-list/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @remirror/extension-list
 
+## 1.2.20
+
+> 2022-03-25
+
+### Patch Changes
+
+- Add an `isViewEditable` helper to determine if the view content is editable.
+
+  Expose the return type of `onAppendTransaction`
+
+- Updated dependencies []:
+  - @remirror/core@1.4.1
+  - @remirror/extension-events@1.0.19
+
 ## 1.2.19
 
 > 2022-03-17