+++ packages/remirror__extension-react-tables/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @remirror/extension-react-tables
 
+## 1.0.29
+
+> 2022-03-25
+
+### Patch Changes
+
+- Add an `isViewEditable` helper to determine if the view content is editable.
+
+  Expose the return type of `onAppendTransaction`
+
+- Updated dependencies []:
+  - @remirror/core@1.4.1
+  - @remirror/extension-positioner@1.2.2
+  - @remirror/extension-tables@1.0.21
+  - @remirror/preset-core@1.0.24
+  - @remirror/react-components@1.0.28
+  - @remirror/react-core@1.1.1
+  - @remirror/react-hooks@1.0.28
+
 ## 1.0.28
 
 > 2022-03-17